+++ Studdype_Final/src/main/java/com/studdype/test/model/biz/board/BookBiz.java
@@ -17,4 +17,5 @@ public interface BookBiz {
 	public Map<Integer, MemberDto> getBookWriterName(int mem_no); // 해당 도서 아이디, 이름 찾아오기
 	public int insertRegisterBook(BookDto dto);	// 작성한 도서 정보 insert
 	public BookDto deleteBook(BookDto dto); // 도서 게시글 삭제
+	public int bookBoardUpdateBook(BookDto dto);	// 도서 게시글 수정
 }
