+++ Studdype_Final/src/main/java/com/studdype/test/model/biz/board/BookBizImpl.java
@@ -74,5 +74,10 @@ public class BookBizImpl implements BookBiz{
 		return resDto;
 	}
 
+	@Override
+	public int bookBoardUpdateBook(BookDto dto) {
+		return bookDao.bookBoardUpdateBook(dto);
+	}
+
 	
 }
